Create tabs for AdditionalDocuments and Focuses
