+++ src/views/EditAry/RightPanel/Methodology/styles.scss
@@ -22,27 +22,6 @@
     background-color: var(--color-foreground);
 }
 
-%content {
-    display: flex;
-    flex-basis: calc(100% / 4);
-    flex-direction: column;
-    flex-grow: 1;
-    overflow: auto;
-    overflow-y: auto;
-
-    :global {
-        .empty {
-            display: block;
-            padding: var(--spacing-small) var(--spacing-medium);
-        }
-    }
-}
-
-%header {
-    font-size: var(--font-size-medium);
-    font-weight: var(--font-weight-light);
-}
-
 .header {
     flex-shrink: 0;
     background-color: var(--color-foreground);
@@ -62,148 +41,17 @@
     background-color: var(--color-foreground);
     overflow: auto;
 
-    .resizable {
+    .top {
         display: flex;
         flex-direction: column;
         width: 100%;
+        height: 100%;
+        min-height: 60%;
         overflow: auto;
 
-        .top {
-            display: flex;
-            flex-direction: column;
-            width: 100%;
-            height: 100%;
-            min-height: 60%;
-            overflow: auto;
-
-            .toptop {
-                height: 50%;
-                min-height: 20%;
-                overflow: auto;
-
-                >.header {
-                    flex-shrink: 0;
-                }
-
-                .attributes-section {
-                    display: flex;
-                    flex-direction: column;
-
-                    >.header {
-                        flex-shrink: 0;
-                    }
-
-                    .scroll-wrap {
-                        flex-grow: 1;
-
-                        .attributes {
-                            display: table;
-                            padding: 0;
-                            width: 100%;
-                            overflow: auto;
-
-                            >.header {
-                                display: table-row;
-
-                                .title {
-                                    @extend %cell;
-                                    @extend %heading;
-                                    text-transform: capitalize;
-                                    color: var(--color-accent);
-                                    font-weight: var(--font-weight-bold);
-                                }
-
-                                .action-buttons {
-                                    @extend %action-buttons;
-                                    @extend %heading;
-                                }
-                            }
-
-                            .row {
-                                display: table-row;
-
-                                .cell {
-                                    @extend %cell;
-                                }
-
-                                .action-buttons {
-                                    @extend %action-buttons;
-                                }
-
-                                :global {
-                                    .widget {
-                                        margin: var(--spacing-small-alt) 0;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            .topbottom {
-                flex-grow: 1;
-                height: 50%;
-                min-height: 40px;
-                overflow: auto;
-
-                .middle-section {
-                    display: flex;
-                    flex-grow: 1;
-                    border-top: var(--width-separator-thin) solid var(--color-separator);
-                    height: 100%;
-
-                    .focuses,
-                    .sectors {
-                        display: flex;
-                        flex-direction: column;
-                        padding: 0 var(--spacing-medium);
-                        height: auto;
-                        overflow-y: auto;
-
-                        .list-selection {
-                            overflow-y: auto;
-                        }
-
-                        :global .list-view {
-                            display: flex;
-                            flex-wrap: wrap;
-
-                            >* {
-                                flex-basis: 50%;
-                            }
-                        }
-
-                    }
-
-                    .location-selection,
-                    .affected-groups {
-                        :global .select-input-with-list {
-                            padding: var(--spacing-small);
-                        }
-                    }
-
-                    .middle-section-item,
-                    .location-selection,
-                    .affected-groups {
-                        @extend %content;
-
-                        .title {
-                            flex-shrink: 0;
-                            padding: var(--spacing-small) var(--spacing-medium);
-                            text-transform: uppercase;
-                            font-size: var(--font-size-large);
-                            font-weight: var(--font-weight-light);
-                        }
-                    }
-                }
-            }
-        }
-
-        .bottom {
-            display: flex;
-            flex-direction: column;
-            height: 30%;
+        .toptop {
+            flex-grow: 1;
+            height: 50%;
             min-height: 40px;
             overflow: auto;
 
@@ -211,7 +59,6 @@
                 display: flex;
                 flex-direction: column;
                 flex-grow: 1;
-                border-top: var(--width-separator-thin) solid var(--color-separator);
                 height: 100%;
 
                 .content {
@@ -245,7 +92,72 @@
             }
         }
 
-    }
+        .topbottom {
+            border-top: var(--width-separator-thin) solid var(--color-separator);
+            height: 50%;
+            min-height: 20%;
+            overflow: auto;
 
+            >.header {
+                flex-shrink: 0;
+            }
+
+            .attributes-section {
+                display: flex;
+                flex-direction: column;
+
+                >.header {
+                    flex-shrink: 0;
+                }
+
+                .scroll-wrap {
+                    flex-grow: 1;
+
+                    .attributes {
+                        display: table;
+                        padding: 0;
+                        width: 100%;
+                        overflow: auto;
+
+                        >.header {
+                            display: table-row;
+
+                            .title {
+                                @extend %cell;
+                                @extend %heading;
+                                text-transform: capitalize;
+                                color: var(--color-accent);
+                                font-weight: var(--font-weight-bold);
+                            }
+
+                            .action-buttons {
+                                @extend %action-buttons;
+                                @extend %heading;
+                            }
+                        }
+
+                        .row {
+                            display: table-row;
+
+                            .cell {
+                                @extend %cell;
+                            }
+
+                            .action-buttons {
+                                @extend %action-buttons;
+                            }
+
+                            :global {
+                                .widget {
+                                    margin: var(--spacing-small-alt) 0;
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+    }
 }
